Refactor integration tests to own test class
Since these tests require special configuration, they were previously setting
up and tearing down a server needlessly. In this dedicated test class, they
may specify their specific configuration and setting up the server as part
of the test instead.
